requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.0
ez-setup==0.9
fakeredis==1.7.1
fastapi==0.75.2
fastapi-mail==1.0.8
fastapi-pagination==0.9.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.3.0
httpx==0.22.0
idna==3.3
Jinja2==3.0.3
limits==1.6
loguru==0.6.0
lxml==4.8.0
MarkupSafe==2.1.0
mysqlclient==2.1.0
packaging==21.3
passlib==1.7.4
pdfkit==1.0.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
PyMySQL==1.0.2
pyparsing==3.0.8
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==3.5.3
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
slowapi==0.1.5
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.5
watchgod==0.7
websockets==10.2
win32-setctime==1.1.0
wrapt==1.14.0
xmltodict==0.13.0